change sidebar to navbar
