--- src/style/scss/Sidenav.scss
@@ -1,9 +1,0 @@
-.logoImg {
-    max-width: 100%;
-}
-
-.subheader, .collapsible-header {
-
-font-size: 18px !important;
-}
-
