+++ src/style/scss/index.scss
@@ -6,24 +6,20 @@ body {
   font-family: "Roboto", sans-serif;
 }
 
-  header, main, footer {
-      padding-left: 300px;
-    }
+  
 .nav-wrapper {
- padding-left: 240px;
+padding-left: 30px;
   max-width: 100%;
 display: flex;
 justify-content: space-between;
 padding-right: 30px;
+
+.logoBox{
+  padding-right: 100px;
+}
 }
     
-    @media only screen and (max-width : 992px) {
-      header, main, footer, .nav-wrapper{
-        
-        padding-left: 0;
-
-      }
-    }
+  
   
     .clearfix {clear:both;}
   
@@ -37,7 +33,7 @@ a.active{
 }
 
 .search-area {
-  width: 80%;
+  width: 60%;
 }
 
 .shopping-cart {
